fix: game offset handling
Previously, values would output negative numbers, as
leastDelayedTimeToSend wasn't of type Timestamp; rather, it was a
number. This caused the values to never subtract properly (as number -
undefined will still equal that same number), which can cause stuff to
wait for negative amounts of time, which turns into 0ms, instead of the
0ms and onward that's required to time everything correctly.

JSON.parse and JSON.stringify has been refactored out, as cloning a
number can simply be done by declaring a variable.